% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloworld.R
\name{helloworld}
\alias{helloworld}
\title{helloworld - the simplest possible function}
\usage{
helloworld()
}
\value{
string hello world
}
\description{
helloworld - the simplest possible function
}
\examples{
helloworld()
}
